Fix bugs in POST method for adding rides
-Also fixed some bugs in fetch for get all rides
-Removed unneeded driver.html page
Relates #34
