--- public/layouts/driver.html
@@ -1,81 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <meta http-equiv="X-UA-Compatible" content="ie=edge">
-  <link rel="stylesheet" href="../styles/styles.css">
-  <link href="https://fonts.googleapis.com/css?family=Playfair+Display|Raleway" rel="stylesheet">
-  <!-- <link rel="icon" href=""> -->
-  <meta name="description" content="rideshare app">
-  <meta name="keywords" content="battuta ibn duck rideshare blablacar rome 2 rio travel">
-  <meta name="author" content="John and Obaydah">
-
-  <title>Battuta</title>
-</head>
-
-<body>
-
-  <header>
-    <h1>Battuta</h1>
-  </header>
-
-  <h2>Profile</h2>
-
-  <h2>Add a Ride</h2>
-
-  <form class="" action="/addRide" method="POST">
-
-    <fieldset>
-      <legend>Offer a Ride</legend>
-      <input type="text" name="date">
-      <input type="text" name="time">
-      <input type="text" name="from">
-      <input type="text" name="to">
-      <input type="text" name="price">
-      <input type="submit" name="submit" value="submit">
-      <input type="reset" name="reset" value="reset">
-    </fieldset>
-
-  </form>
-
-
-
-  <table>
-    <caption>My Rides</caption>
-    <tr>
-      <th>Date</th>
-      <th>Time</th>
-      <th>From</th>
-      <th>To</th>
-      <th>Price</th>
-      <th></th>
-      <th></th>
-    </tr>
-    <tr>
-      <td>1 July 2019</td>
-      <td>14:30pm</td>
-      <td>Haifa</td>
-      <td>Jerusalem</td>
-      <td>50</td>
-      <td>Edit</td>
-      <td>Delete</td>
-
-    </tr>
-  </table>
-
-  <h2>My car</h2>
-
-  <footer>Battuta app copyright 2019 J&O Solutions, all rights reserved</footer>
-
-  <script type="text/javascript">
-  let form = document.querySelector('form');
-  form.addEventListener("submit", event => {
-    event.preventDefault();
-    console.log("adding ride:", form.elements.value);
-
-  });
-  </script>
-
-</body>
-</html>
