+++ public/layouts/index.html
@@ -30,7 +30,7 @@
 
   <!-- <a href="/auth_check"><button>Am I Authenticated?</button></a> -->
 
-  <!-- <h2>Profile</h2> -->
+  <h2>Going your way</h2>
 
   <h2></h2>
 
@@ -77,7 +77,7 @@
   </table>
 </fieldset>
 
-  <script type="text/javascript" src="public/scripts/index.js"></script>
+  <!-- <script type="text/javascript" src="public/scripts/index.js"></script> -->
   <script type="text/javascript" src="public/scripts/fetch.js"></script>
   <script type="text/javascript" src="public/scripts/dom.js"></script>
 </body>
