+++ src/queries/getData.js
@@ -3,7 +3,7 @@ const dbConnection = require('../database/db_connection.js');
 const getRides = (callback) => {
   dbConnection.query(`SELECT * FROM rides`, (error, result) => {
     if (error) return callback(error);
-    // console.log('This is the result: ' + result.rows);
+    console.log('This is the result: ' + result.rows);
     callback(null, result);
   })
 }
