+++ src/queries/postData.js
@@ -3,15 +3,13 @@ const getData = require('./getData');
 
 const addRide = (driverid, pickup, dropoff, price, response, cb) => {
   dbConnection.query(
-    // `SELECT * FROM rides;`,
     `INSERT INTO rides (driver_id, pickup, dropoff, price) VALUES ($1, $2, $3, $4);`,
     [driverid, pickup, dropoff, price],
     (err, res) => {
       console.log(res);
-      // console.log(err)
-      // console.log(res.rows)
+      console.log(res.rows)
       if (err) return cb(err);
-      cb(null, res, console.log(response));
+      cb(null, res);
     }
   )
 }
