+++ src/router.js
@@ -27,12 +27,14 @@ const router = (req, res) => {
         res.writeHead(302, {'Location': '/', 'Set-Cookie': 'jwt=0; Max-Age=0'});
         return res.end();
       }
-    } else if (url === '/getOffers') {
-      handler.data(res);
     }
-    else if (url.indexOf('public') !== -1){
-        handler.public(url, res);
-    } else if (url === '/addRide') {
+    // else if (url === '/getOffers') {
+    //   handler.data(res);
+    // }
+    // else if (url.indexOf('public') !== -1){
+    //     handler.public(url, res);
+    // }
+    else if (url === '/addRide') {
       if (req.method === 'POST') {
         handler.post(req, res);
         res.writeHead(302, {'Location': '/', "Content-Type": 'text/plain'});
